--- app/src/main/java/com/example/beelocal/domain/FirestoreCollections.kt
@@ -1,10 +1,0 @@
-package com.example.beelocal.domain
-
-enum class FirestoreCollections(val value: String) {
-    USERS("users"),
-    DAILY_CHALLENGES("daily_challenges"),
-    FEED("feed"),
-    ROUTES("routes"),
-    REVIEWS("reviews"),
-    BINGO_CARDS("bingo_cards")
-}
