--- app/src/main/java/com/example/beelocal/model/BingoCard.kt
@@ -1,11 +1,0 @@
-package com.example.beelocal.model
-
-import com.google.firebase.Timestamp
-import com.google.firebase.firestore.DocumentId
-
-data class BingoCard(
-    @DocumentId
-    val id: String = "",
-    val weekStartDate: Timestamp = Timestamp.now(),
-    val tasks: List<BingoTask> = emptyList()
-)
