--- app/src/main/java/com/example/beelocal/model/BingoTask.kt
@@ -1,10 +1,0 @@
-package com.example.beelocal.model
-
-import com.example.beelocal.model.types.BingoTaskType
-
-data class BingoTask(
-    val id: String = "",
-    val title: String = "",
-    val description: String = "",
-    val type: BingoTaskType = BingoTaskType.PHOTO
-)
