--- app/src/main/java/com/example/beelocal/model/NotificationSettings.kt
@@ -1,6 +1,0 @@
-package com.example.beelocal.model
-
-data class NotificationSettings(
-    val emailEnabled: Boolean = true,
-    val phoneEnabled: Boolean = true
-)
