--- app/src/main/java/com/example/beelocal/model/RoutePoint.kt
@@ -1,11 +1,0 @@
-package com.example.beelocal.model
-
-import com.google.firebase.firestore.GeoPoint
-
-data class RoutePoint(
-    val name: String = "",
-    val description: String = "",
-    val location: GeoPoint = GeoPoint(0.0, 0.0),
-    val quizQuestion: String = "",
-    val quizAnswer: String = ""
-)
