--- app/src/main/java/com/example/beelocal/model/RouteReview.kt
@@ -1,13 +1,0 @@
-package com.example.beelocal.model
-
-import com.google.firebase.firestore.DocumentId
-
-data class RouteReview(
-    @DocumentId
-    val id: String = "",
-    val userId: String = "",
-    val username: String = "",
-    val rating: Int = 0,
-    val comment: String = "",
-    val photoUrls: List<String> = emptyList()
-)
