--- app/src/main/java/com/example/beelocal/model/UserBingoProgress.kt
@@ -1,11 +1,0 @@
-package com.example.beelocal.model
-
-import com.google.firebase.firestore.DocumentId
-
-data class UserBingoProgress(
-    @DocumentId
-    val id: String = "",
-    val userId: String = "",
-    val bingoCardId: String = "",
-    val completedTaskIds: List<String> = emptyList()
-)
