--- app/src/main/java/com/example/beelocal/model/types/BingoTaskType.kt
@@ -1,6 +1,0 @@
-package com.example.beelocal.model.types
-
-enum class BingoTaskType {
-    PHOTO,
-    LOCATION
-}
