--- app/src/main/java/com/example/beelocal/model/types/FeedEntryType.kt
@@ -1,7 +1,0 @@
-package com.example.beelocal.model.types
-
-enum class FeedEntryType {
-    DAILY_CHALLENGE,
-    ROUTE_COMPLETED,
-    BINGO_TASK_COMPLETED
-}
