+++ app/src/main/java/com/pv239/beelocal/domain/FirestoreCollections.kt
@@ -1,0 +1,10 @@
+package com.pv239.beelocal.domain
+
+enum class FirestoreCollections(val value: String) {
+    USERS("users"),
+    DAILY_CHALLENGES("daily_challenges"),
+    FEED("feed"),
+    ROUTES("routes"),
+    REVIEWS("reviews"),
+    BINGO_CARDS("bingo_cards")
+}
