+++ app/src/main/java/com/pv239/beelocal/model/BingoCard.kt
@@ -1,0 +1,11 @@
+package com.pv239.beelocal.model
+
+import com.google.firebase.Timestamp
+import com.google.firebase.firestore.DocumentId
+
+data class BingoCard(
+    @DocumentId
+    val id: String = "",
+    val weekStartDate: Timestamp = Timestamp.now(),
+    val tasks: List<BingoTask> = emptyList()
+)
