+++ app/src/main/java/com/pv239/beelocal/model/NotificationSettings.kt
@@ -1,0 +1,6 @@
+package com.pv239.beelocal.model
+
+data class NotificationSettings(
+    val emailEnabled: Boolean = true,
+    val phoneEnabled: Boolean = true
+)
