+++ app/src/main/java/com/pv239/beelocal/model/RoutePoint.kt
@@ -1,0 +1,11 @@
+package com.pv239.beelocal.model
+
+import com.google.firebase.firestore.GeoPoint
+
+data class RoutePoint(
+    val name: String = "",
+    val description: String = "",
+    val location: GeoPoint = GeoPoint(0.0, 0.0),
+    val quizQuestion: String = "",
+    val quizAnswer: String = ""
+)
