+++ app/src/main/java/com/pv239/beelocal/model/RouteReview.kt
@@ -1,0 +1,13 @@
+package com.pv239.beelocal.model
+
+import com.google.firebase.firestore.DocumentId
+
+data class RouteReview(
+    @DocumentId
+    val id: String = "",
+    val userId: String = "",
+    val username: String = "",
+    val rating: Int = 0,
+    val comment: String = "",
+    val photoUrls: List<String> = emptyList()
+)
