+++ app/src/main/java/com/pv239/beelocal/model/types/BingoTaskType.kt
@@ -1,0 +1,6 @@
+package com.pv239.beelocal.model.types
+
+enum class BingoTaskType {
+    PHOTO,
+    LOCATION
+}
