+++ app/src/main/java/com/pv239/beelocal/model/types/FeedEntryType.kt
@@ -1,0 +1,7 @@
+package com.pv239.beelocal.model.types
+
+enum class FeedEntryType {
+    DAILY_CHALLENGE,
+    ROUTE_COMPLETED,
+    BINGO_TASK_COMPLETED
+}
